requirements.txt
# Core Dependencies
asyncio-mqtt==0.16.2
aiohttp==3.9.1
websockets==12.0
pexpect==4.9.0
psutil==5.9.6

# AI/ML
openai==1.3.0
anthropic==0.7.0
transformers==4.35.0
langchain==0.0.340
chromadb==0.4.18
sentence-transformers==2.2.2

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Web & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
rich==13.7.0
click==8.1.7
PyYAML==6.0.1

# Parsing & Processing
beautifulsoup4==4.12.2
feedparser==6.0.10

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For advanced features
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2